東大とソフトバンクのメニューが逆だった不具合の修正
